fix(player): preserve diagnostic counter types

+++ composeApp/src/androidMain/kotlin/com/yfuse/feature/player/YCorePlayerRuntime.android.kt
@@ -156,8 +156,8 @@ private fun createYCorePlaybackSession(
     performanceMemory: PlaybackPerformanceMemory,
     startedAtEpochMs: Long,
     initialPositionMs: Long,
-    initialBufferEvents: Long,
-    initialDroppedFrames: Long,
+    initialBufferEvents: Int,
+    initialDroppedFrames: Int,
 ): YCorePlaybackSession =
     YCorePlaybackSession(
         engine = engine,
